refactor(habitat): ratchet recipe stage topology
Replace the Standard-specific stage topology parser with generic recipe-stage and recipe-step blueprint authority backed by structure.toml. Relocate shared family authoring surfaces to public.config.ts, return the Swooper domain-root import law to niche ownership, and refresh the canonical authority ledger and execution-surface reports.\n\nRuntime configuration and stage ordering are unchanged; focused Habitat rules, Swooper/Habitat typecheck-test-build proof, strict OpenSpec validation, generated-currentness, and fresh permanent review roles pass. Aggregate policy proof follows the separate redundant-scan performance repair.\n\nReferences: docs/projects/engine-refactor-v1/architecture-normalization-packet.md
