+++ .habitat/blueprints/recipe-step/require_recipe_step_source_topology/rule.json
@@ -19,7 +19,10 @@
   "pathCoverage": [
     {
       "kind": "exact-path",
-      "patterns": ["mods/*/src/recipes/*/stages/*/steps", "mods/*/src/recipes/*/stages/*/steps/**"]
+      "patterns": [
+        "mods/*/src/recipes/*/stages/**/steps",
+        "mods/*/src/recipes/*/stages/**/steps/**"
+      ]
     }
   ],
   "supportFiles": {
