+++ .habitat/blueprints/recipe-step/require_recipe_step_source_topology/structure.toml
@@ -2,14 +2,14 @@ schemaVersion = 1
 
 [[scopes]]
 name = "recipe-step-roots"
-root = "mods/*/src/recipes/*/stages/*/steps"
+root = "mods/*/src/recipes/*/stages/**/steps"
 kind = "directory"
 mode = "closed"
 allowed = ["+([a-z0-9])*(-+([a-z0-9]))"]
 
 [[scopes]]
 name = "recipe-step-containers"
-root = "mods/*/src/recipes/*/stages/*/steps/+([a-z0-9])*(-+([a-z0-9]))"
+root = "mods/*/src/recipes/*/stages/**/steps/+([a-z0-9])*(-+([a-z0-9]))"
 kind = "directory"
 mode = "open"
 required = ["config.ts", "step.ts"]
@@ -17,12 +17,12 @@ forbidden = ["contract.ts", "index.ts", "*.contract.ts"]
 
 [[scopes]]
 name = "recipe-step-config-modules"
-root = "mods/*/src/recipes/*/stages/*/steps/+([a-z0-9])*(-+([a-z0-9]))/config.ts"
+root = "mods/*/src/recipes/*/stages/**/steps/+([a-z0-9])*(-+([a-z0-9]))/config.ts"
 kind = "file"
 mode = "open"
 
 [[scopes]]
 name = "recipe-step-runtime-modules"
-root = "mods/*/src/recipes/*/stages/*/steps/+([a-z0-9])*(-+([a-z0-9]))/step.ts"
+root = "mods/*/src/recipes/*/stages/**/steps/+([a-z0-9])*(-+([a-z0-9]))/step.ts"
 kind = "file"
 mode = "open"
