+++ .habitat/civ7/mapgen/domains/ecology/rules/require_public_ecology_surfaces_and_retired_topology_removal/rule.json
@@ -13,7 +13,7 @@
   "ownerProject": "mod-swooper-maps",
   "lane": "enforced",
   "forbids": "active ecology recipe stages importing/exporting ecology ops or rules internals",
-  "why": "Ecology recipe stages compose public domain surfaces; retired wrapper topology is enforced by preserve_standard_stage_topology_and_path_invariants.",
+  "why": "Ecology recipe stages compose public domain surfaces; the generic recipe-stage blueprint owns their legal source topology.",
   "remediate": null,
   "message": "Use public ecology domain surfaces and keep retired ecology topology out of active source.",
   "pathCoverage": [
