--- .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/preserve_standard_stage_topology_and_path_invariants/rule.json
@@ -1,40 +1,0 @@
-{
-  "schemaVersion": 2,
-  "id": "preserve_standard_stage_topology_and_path_invariants",
-  "title": "Preserve Standard Stage Topology And Path Invariants",
-  "placement": {
-    "niche": "civ7/mapgen/pipeline/swooper-maps-standard-recipe",
-    "blueprint": "_self",
-    "category": "structure"
-  },
-  "operation": {
-    "kind": "check"
-  },
-  "ownerProject": "mod-swooper-maps",
-  "lane": "enforced",
-  "forbids": "standard recipe stage directory drift from the contract manifest/runtime recipe, missing stage entrypoints, or retired stage directories",
-  "why": "The standard MapGen pipeline file tree must stay closed around active stages derived from the contract manifest and runtime recipe; source-order and step parity checks are owned by separate rules.",
-  "remediate": null,
-  "message": "Update the standard contract manifest, runtime recipe, and stage-root topology together; shared support hubs are allowed but are not active stage ids.",
-  "pathCoverage": [
-    {
-      "kind": "exact-path",
-      "patterns": [
-        "mods/mod-swooper-maps/src/recipes/standard/contract-manifest.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/recipe.ts",
-        "mods/mod-swooper-maps/src/recipes/standard/stages/*"
-      ]
-    }
-  ],
-  "supportFiles": {
-    "baseline": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/preserve_standard_stage_topology_and_path_invariants/baseline.json"
-  },
-  "runner": {
-    "name": "habitat",
-    "mode": "script",
-    "runtime": "node",
-    "files": {
-      "script": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/preserve_standard_stage_topology_and_path_invariants/check.mjs"
-    }
-  }
-}
