+++ .habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/require_domain_contract_roots_in_step_contracts/rule.json
@@ -1,0 +1,37 @@
+{
+  "schemaVersion": 2,
+  "id": "require_domain_contract_roots_in_step_contracts",
+  "title": "Require Domain Contract Roots In Step Contracts",
+  "placement": {
+    "niche": "civ7/mapgen/pipeline/swooper-maps-standard-recipe",
+    "blueprint": "_self",
+    "category": "boundary"
+  },
+  "operation": {
+    "kind": "check"
+  },
+  "ownerProject": "mod-swooper-maps",
+  "lane": "enforced",
+  "forbids": "Swooper Standard recipe step config modules deep-importing the mod-owned @mapgen/domain implementation",
+  "why": "The Swooper Standard recipe binds StepContract definitions to each domain's public root rather than the mod's private operation or model files.",
+  "remediate": null,
+  "message": "Swooper Standard StepContract config modules must import only from @mapgen/domain/<domain>.",
+  "pathCoverage": [
+    {
+      "kind": "exact-path",
+      "patterns": ["mods/mod-swooper-maps/src/recipes/standard/stages/**/steps/*/config.ts"]
+    }
+  ],
+  "scanRoots": ["mods/mod-swooper-maps/src/recipes/standard/stages"],
+  "hookCheck": true,
+  "supportFiles": {
+    "baseline": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/require_domain_contract_roots_in_step_contracts/baseline.json"
+  },
+  "runner": {
+    "name": "grit",
+    "files": {
+      "pattern": ".habitat/civ7/mapgen/pipeline/swooper-maps-standard-recipe/rules/require_domain_contract_roots_in_step_contracts/pattern.md"
+    },
+    "patternName": "require_domain_contract_roots_in_step_contracts"
+  }
+}
