+++ mods/mod-swooper-maps/src/recipes/standard/stages/ecology/public.config.ts
