+++ mods/mod-swooper-maps/src/recipes/standard/stages/foundation/public.config.ts
