+++ mods/mod-swooper-maps/src/recipes/standard/stages/hydrology-climate-baseline/index.ts
@@ -1,6 +1,6 @@
 import { createStage, Type } from "@swooper/mapgen-core/authoring";
 import { orderStandardStageSteps } from "../../contract-manifest.js";
-import { HydrologyClimateBaselinePublicSchema } from "../hydrology-public-config.js";
+import { HydrologyClimateBaselinePublicSchema } from "../hydrology/public.config.js";
 import { ClimateBaselineStep } from "./steps/climate-baseline/step.js";
 
 const HydrologyDrynessKnobSchema = Type.Union(
