+++ mods/mod-swooper-maps/src/recipes/standard/stages/hydrology/public.config.ts
