+++ mods/mod-swooper-maps/src/recipes/standard/stages/placement/public.config.ts
